pyproject.toml
[project]
name = "smart-fit"
version = "0.1.0"
description = ""
authors = [
    {name = "amankishore32",email = "aman.kishore32@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai (>=0.8.5,<0.9.0)",
    "numpy (>=2.3.5,<3.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pypdf (>=6.4.1,<7.0.0)",
    "fastapi (>=0.124.0,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)"
]

[tool.poetry]
packages = [{include = "smart_fit", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
